// apps/web/lib/api-types/canonical-reader.generated.ts
export type Align = ("start" | "left" | "right" | "center" | "justify") | null;
export type Bold = boolean | null;
export type Family = ("serif" | "sans-serif" | "monospace") | null;
export type Id14 = string;
export type IndentEm = number | null;
export type Italic = boolean | null;
export type LineHeight = number | null;
export type RelativeSize = number | null;
export type SmallCaps = boolean | null;
export type SpaceAfterEm = number | null;
export type SpaceBeforeEm = number | null;
export type VerticalAlign = ("baseline" | "super" | "sub") | null;
// ...
export interface Style {
  align?: Align;
  bold?: Bold;
  family?: Family;
  id: Id14;
  indent_em?: IndentEm;
  italic?: Italic;
  line_height?: LineHeight;
  relative_size?: RelativeSize;
  small_caps?: SmallCaps;
  space_after_em?: SpaceAfterEm;
  space_before_em?: SpaceBeforeEm;
  vertical_align?: VerticalAlign;
}